Améliore la relecture du code
