+++ src/views/Simulation/Recapitulatif.vue
@@ -60,202 +60,14 @@
 
 <script>
 import Actions from "@/components/Actions"
-import IndividuQuestions from "@/lib/IndividuQuestions"
 import {
   capitalize,
-  displayEnumValue,
   displayValue,
-  displayYesNoValue,
   executeFunctionOrReturnValue,
-  displayCurrencyValue,
-  displayDepcomValue,
 } from "@/lib/Utils"
-import Individu from "@/lib/Individu"
-import FamilleQuestions from "@/lib/FamilleQuestions"
-import ParentQuestions from "@/lib/ParentQuestions"
-import Ressource from "@/lib/Ressource"
-import { ressourceCategories, ressourceTypes } from "@/constants/resources"
-import Logement from "@/lib/Logement"
+import { SIMPLE_STEPS, COMPLEX_STEPS } from "@/lib/Recapitulatif"
+import { ENTITIES_PROPERTIES } from "@/lib/State/steps"
 
-function getIndividuByStep(step, component) {
-  const role = step.id.split("_")[0]
-  return Individu.get(
-    component.$store.getters.peopleParentsFirst,
-    role,
-    step.id,
-    component.$store.state.dates
-  ).individu
-}
-
-const SIMPLE_STEPS = {
-  enfant_a_charge(step) {
-    const individu = getIndividuByStep(step, this)
-    return [
-      {
-        label: capitalize(
-          individu._role === "demandeur"
-            ? `Figurez-vous sur la dernière déclaration d'impôts de vos parents ?`
-            : `${Individu.label(
-                individu,
-                "nom"
-              )} figure-t-il/elle sur votre dernière déclaration d'impôts sur le revenu ?`
-        ),
-        value: displayYesNoValue(
-          individu["enfant_a_charge"][this.$store.state.dates.thisYear.id]
-        ),
-      },
-    ]
-  },
-  statut_marital(step) {
-    const individu = getIndividuByStep(step, this)
-    return [
-      {
-        label: "Quelle est votre relation avec votre conjoint ?",
-        value: displayEnumValue(
-          individu[step.variable],
-          Individu.situationsFamiliales
-        ),
-      },
-    ]
-  },
-  _firstName(step) {
-    return [
-      {
-        label: "Quel est le prénom de votre enfant ?",
-        value: getIndividuByStep(step, this)[step.variable],
-      },
-    ]
-  },
-  _bourseCriteresSociauxCommuneDomicileFamilial() {
-    return [
-      {
-        label: "Quel est le code postal de la commune de vos parents ?",
-        value: displayDepcomValue(
-          this.individu._bourseCriteresSociauxCommuneDomicileFamilialCodePostal,
-          this.individu._bourseCriteresSociauxCommuneDomicileFamilialNomCommune
-        ),
-      },
-    ]
-  },
-  depcom() {
-    const menage = this.$store.getters.getMenage
-    return [
-      {
-        label: "Quel est votre code postal ?",
-        value: displayDepcomValue(menage._codePostal, menage._nomCommune),
-      },
-    ]
-  },
-  coloc() {
-    return [
-      {
-        label: "Est-ce une colocation ?",
-        value: displayYesNoValue(this.$store.getters.getMenage.coloc),
-      },
-    ]
-  },
-  logement_chambre() {
-    return [
-      {
-        label: `Est-ce une chambre ? <span class="help">Une chambre est un logement qui ne comporte qu'une seule pièce et n'est pas équipé d'un WC.</span>`,
-        value: displayYesNoValue(
-          this.$store.getters.getMenage.logement_chambre
-        ),
-      },
-    ]
-  },
-  participation_frais() {
-    return [
-      {
-        label: `Participez-vous aux frais du logement ?<span class="help">Par exemple aux dépenses d'électricité, de téléphone, etc.</span>`,
-        value: displayYesNoValue(
-          this.$store.getters.getMenage.participation_frais
-        ),
-      },
-    ]
-  },
-}
-
-const COMPLEXE_STEPS = {
-  "ressources/montants": {
-    matcher(step) {
-      return step.key.match(/ressources\/montants\/(\w)*/)
-    },
-    fn(step) {
-      const key_split = step.key.split("/")
-      const id = key_split[1]
-      const individu = getIndividuByStep({ id, role: id.split("_")[0] }, this)
-
-      const categoryId = key_split[key_split.length - 1]
-      const ressources = Ressource.getIndividuRessourceTypes(
-        individu,
-        this.$store.state.situation
-      )
-      const category = ressourceCategories.find(
-        (category) => category.id === categoryId
-      )
-      const result = [
-        {
-          isChapterSubtitle: true,
-          label: category && capitalize(category.label(individu)),
-          value: "",
-        },
-        ...ressourceTypes
-          .filter((type) => type.category === categoryId && ressources[type.id])
-          .map((type) => {
-            return {
-              label: capitalize(type.label),
-              value: Object.entries(individu[type.id]).reduce(
-                (accum, [key, value]) => {
-                  accum[key] = displayCurrencyValue(value)
-                  return accum
-                },
-                {}
-              ),
-            }
-          }),
-      ]
-      return result
-    },
-  },
-  logement: {
-    matcher(step) {
-      return step.key.match(/\/logement$/)
-    },
-    fn() {
-      const menage = this.$store.state.situation.menage
-      return [
-        {
-          label: "Êtes-vous ?",
-          value: Logement.getStatutOccupationLabel(
-            menage.statut_occupation_logement
-          ),
-        },
-      ]
-    },
-  },
-  loyer: {
-    matcher(step) {
-      return step.key.match(/\/loyer$/)
-    },
-    fn() {
-      const loyerData = Logement.getLoyerData(
-        this.$store.getters.getMenage,
-        this.$store.getters.getLogementStatut || ""
-      )
-      return [
-        {
-          label: loyerData.loyerQuestion.label,
-          value: displayCurrencyValue(loyerData.loyerQuestion.selectedValue),
-        },
-        loyerData.chargesQuestion && {
-          label: loyerData.chargesQuestion.label,
-          value: displayCurrencyValue(loyerData.chargesQuestion.selectedValue),
-        },
-      ].filter((item) => item)
-    },
-  },
-}
 export default {
   name: "Recapitulatif",
   components: {
@@ -296,64 +108,44 @@ export default {
       this.$push()
     },
 
-    buildMutualizedQuestion(data) {
-      return data.question
+    buildMutualizedQuestion({ question, value, component }) {
+      return question
         ? [
             {
               label: capitalize(
-                executeFunctionOrReturnValue(
-                  data.question,
-                  "question",
-                  data.component
-                )
+                executeFunctionOrReturnValue(question, "question", component)
               ),
-              value: displayValue(
-                data.individu[data.variable],
-                data.question,
-                data.component
-              ),
+              value: displayValue(value, question, component),
             },
           ]
         : []
     },
 
     questionsPerStep(step) {
-      if (!step.entity || !step.variable) console.log(step)
-      let data = {
-        variable: step.variable,
-      }
       if (SIMPLE_STEPS[step.variable]) {
         return SIMPLE_STEPS[step.variable].bind(this)(step)
       }
       if (step.variable === undefined) {
-        const match = Object.keys(COMPLEXE_STEPS).find((key) =>
-          COMPLEXE_STEPS[key].matcher(step)
+        const match = Object.keys(COMPLEX_STEPS).find((key) =>
+          COMPLEX_STEPS[key].matcher(step)
         )
         if (match) {
-          return COMPLEXE_STEPS[match].fn.bind(this)(step)
+          return COMPLEX_STEPS[match].fn.bind(this)(step)
         }
       }
 
-      switch (step.entity) {
-        case "individu":
-          data.individu = getIndividuByStep(step, this)
-
-          data.component = { ...this, individu: data.individu }
-          data.question = IndividuQuestions[step.variable]
-          return this.buildMutualizedQuestion(data)
-        case "famille":
-          data.question = FamilleQuestions[step.variable]
-          data.component = this
-          return this.buildMutualizedQuestion(data)
-
-        case "parents":
-          data.question = ParentQuestions[step.variable]
-          data.component = this
-
-          return this.buildMutualizedQuestion(data)
-        default:
-          console.log("### This step is not displayed:", step)
-          break
+      if (ENTITIES_PROPERTIES[step.entity]) {
+        const entity = ENTITIES_PROPERTIES[step.entity].loadEntity({
+          ...this,
+          params: step,
+        })
+        return this.buildMutualizedQuestion({
+          question: ENTITIES_PROPERTIES[step.entity].STEPS[step.variable],
+          value: entity[step.variable],
+          component: { ...this, entity },
+        })
+      } else {
+        console.log("### This step is not displayed:", step)
       }
       return []
     },
